move user test base class.
